fix(franchise): move read_boot_id to core._linux_proc to fix L2->L1 contract violation
Adds core/_linux_proc.py with stdlib-only /proc readers (read_boot_id,
read_starttime_ticks), re-exports from execution/linux_tracing.py to
preserve existing callers, and updates franchise/_api.py to import from
core._linux_proc instead of execution.linux_tracing.

Also removes unused variable original_mark in test_signal_guard.py.
